+++ gatsby-node.js
@@ -1,5 +1,5 @@
 const { validate } = require('superstruct');
-const { createIndex, listDocuments, setMapping } = require('./src/elasticsearch');
+const { createIndex, listDocuments, setMapping, getMapping } = require('./src/elasticsearch');
 const { checkDocument, checkNode } = require('./src/matcher');
 const { OptionsStruct } = require('./src/types');
 
@@ -27,7 +27,17 @@ exports.createPagesStatefully = async ({ graphql, reporter }, rawOptions) => {
     if (options.provider !== 'elastic.co') {
       // elastic.co just uses 'documents' and doesn't accept other indexes/mappings
       await createIndex(options);
-      await setMapping(options);
+      const response = await getMapping(options);
+      const existingIndexes = response[options.index].mappings.properties;
+
+      await setMapping({
+        ...options,
+        mapping: Object.fromEntries(
+          Object.entries(options.mapping).filter(([key, _]) => {
+            return !Object.keys(existingIndexes).includes(key);
+          })
+        )
+      });
     }
 
     const nodes = options.selector(data).map((node) => options.toDocument(node));